[tests] move EntityObserverTests tests to ObserverRegistry
